fix: Caixa models, controller, BLL; Fix AreaArmazenagem, CaixaLeitura, Pallet, Equipamento (Models)

+++ SIAG-CRATO/BLLs/AreaArmzenagem/AreaArmazenagemBLL.cs
@@ -19,7 +19,7 @@ public class AreaArmazenagemBLL
         return areasArmazenagem.ToList();
     }
 
-    public static async Task<AreaArmazenagemModel?> GetByIdAsync(int id)
+    public static async Task<AreaArmazenagemModel?> GetByIdAsync(long id)
     {
         var sql = $@"{AreaArmazenagemQuery.SELECT} WHERE id_areaarmazenagem = @id";
